# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "abacus",
    "description": "",
    "license": {
      "name": "EUPL-1.2",
      "identifier": "EUPL-1.2"
    },
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/elections/{election_id}/committee_sessions": {
      "post": {
        "summary": "Create a new [CommitteeSession]. (coordinator_gsb)",
        "operationId": "committee_session_create",
        "parameters": [
          {
            "name": "election_id",
            "in": "path",
            "description": "Election database id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ElectionId"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Committee session created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommitteeSession"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Request cannot be completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie_auth": [
              "coordinator_gsb"
            ]
          }
        ]
      }
    },
    "/api/elections/{election_id}/committee_sessions/{committee_session_id}": {
      "put": {
        "summary": "Update a [CommitteeSession]. (coordinator_gsb)",
        "operationId": "committee_session_update",
        "parameters": [
          {
            "name": "election_id",
            "in": "path",
            "description": "Election database id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ElectionId"
            }
          },
          {
            "name": "committee_session_id",
            "in": "path",
            "description": "Committee session database id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CommitteeSessionId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommitteeSessionUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Committee session updated successfully"
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Committee session not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie_auth": [
              "coordinator_gsb"
            ]
          }
        ]
      },
      "delete": {
        "summary": "Delete a [CommitteeSession]. (coordinator_gsb)",
        "operationId": "committee_session_delete",
        "parameters": [
          {
            "name": "election_id",
            "in": "path",
            "description": "Election database id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ElectionId"
            }
          },
          {
            "name": "committee_session_id",
            "in": "path",
            "description": "Committee session database id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CommitteeSessionId"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Committee session deleted successfully"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Committee session not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Request cannot be completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie_auth": [
              "coordinator_gsb"
            ]
          }
        ]
      }
    },
...
    "/api/elections/{election_id}/committee_sessions/{committee_session_id}/investigations": {
      "get": {
        "summary": "Get a list of all [crate::domain::investigation::PollingStationInvestigation]s for a committee session (coordinator_gsb)",
        "operationId": "committee_session_investigations",
        "parameters": [
          {
            "name": "election_id",
            "in": "path",
            "description": "Election database id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ElectionId"
            }
          },
          {
            "name": "committee_session_id",
            "in": "path",
            "description": "Committee session database id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CommitteeSessionId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Investigation listing successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Committee session not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie_auth": [
              "coordinator_gsb"
            ]
          }
        ]
      }
    },
    "/api/elections/{election_id}/committee_sessions/{committee_session_id}/status": {
      "put": {
        "summary": "Change the status of a [CommitteeSession]. (coordinator_gsb)",
        "operationId": "committee_session_status_change",
        "parameters": [
          {
            "name": "election_id",
            "in": "path",
            "description": "Election database id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ElectionId"
            }
          },
          {
            "name": "committee_session_id",
            "in": "path",
            "description": "Committee session database id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CommitteeSessionId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommitteeSessionStatusChangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Committee session status changed successfully"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Committee session not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Request cannot be completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie_auth": [
              "coordinator_gsb"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommitteeSession": {
        "type": "object",
        "description": "Committee session",
        "required": [
          "id",
          "number",
          "election_id",
          "location",
          "status"
        ],
        "properties": {
          "election_id": {
            "$ref": "#/components/schemas/ElectionId"
          },
          "id": {
            "$ref": "#/components/schemas/CommitteeSessionId"
          },
          "location": {
            "type": "string"
          },
          "number": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "overview_pdf": {
            "$ref": "#/components/schemas/FileId"
          },
          "results_eml": {
            "$ref": "#/components/schemas/FileId"
          },
          "results_pdf": {
            "$ref": "#/components/schemas/FileId"
          },
          "start_date_time": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/CommitteeSessionStatus"
          }
        },
        "additionalProperties": false
      },
      "CommitteeSessionId": {
        "type": "integer",
        "format": "int32",
        "minimum": 0
      },
      "CommitteeSessionStatus": {
        "type": "string",
        "description": "Committee session status",
        "enum": [
          "created",
          "in_preparation",
          "data_entry",
          "paused",
          "completed"
        ]
      },
      "CommitteeSessionStatusChangeRequest": {
        "type": "object",
        "description": "Committee session status change request",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "$ref": "#/components/schemas/CommitteeSessionStatus"
          }
        },
        "additionalProperties": false
      },
      "CommitteeSessionUpdateRequest": {
        "type": "object",
        "description": "Committee session update request",
        "required": [
          "location",
          "start_date",
          "start_time"
        ],
        "properties": {
          "location": {
            "type": "string"
          },
          "start_date": {
            "type": "string"
          },
          "start_time": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "ElectionId": {
        "type": "integer",
        "format": "int32",
        "minimum": 0
      },
...
      "ErrorReference": {
        "type": "string",
        "description": "Error reference used to show the corresponding error message to the end-user",
        "enum": [
          "AirgapViolation",
          "AlreadyInitialised",
          "ApportionmentAllListsExhausted",
          "ApportionmentCommitteeSessionNotCompleted",
          "ApportionmentDrawingOfLotsRequired",
          "ApportionmentZeroVotesCast",
          "CommitteeSessionPaused",
          "DatabaseError",
          "DataEntryAlreadyClaimed",
          "DataEntryAlreadyFinalised",
          "DataEntryCannotBeDeleted",
          "DataEntryGetNotAllowed",
          "DataEntryNotAllowed",
          "EmlImportError",
          "EmlError",
          "EntryNotFound",
          "EntryNotUnique",
          "Forbidden",
          "InternalServerError",
          "InvalidCommitteeSessionStatus",
          "InvalidData",
          "InvalidHash",
          "InvalidJson",
          "InvalidPassword",
          "InvalidPoliticalGroup",
          "InvalidSession",
          "InvalidStateTransition",
          "InvalidUsernameOrPassword",
          "InvalidVoteCandidate",
          "InvalidVoteGroup",
          "InvalidXml",
          "InvestigationHasDataEntryOrResult",
          "InvestigationRequiresCorrectedResults",
          "NotInitialised",
          "OwnAccountCannotBeDeleted",
          "PasswordRejectionSameAsOld",
          "PasswordRejectionSameAsUsername",
          "PasswordRejectionTooShort",
          "PdfGenerationError",
          "PollingStationRepeated",
          "PollingStationValidationErrors",
          "RequestPayloadTooLarge",
          "Unauthorized",
          "UsernameNotUnique",
          "UserNotFound"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Response structure for errors",
        "required": [
          "error",
          "fatal",
          "reference"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "fatal": {
            "type": "boolean"
          },
          "reference": {
            "$ref": "#/components/schemas/ErrorReference"
          }
        },
        "additionalProperties": false
      },
...
      "FileId": {
        "type": "integer",
        "format": "int32",
        "minimum": 0
      },
...
      "InvestigationListResponse": {
        "type": "object",
        "description": "Investigation list response",
        "required": [
          "investigations"
        ],
        "properties": {
          "investigations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PollingStationInvestigation"
            }
          }
        },
        "additionalProperties": false
      },
...
      "PollingStationId": {
        "type": "integer",
        "format": "int32",
        "minimum": 0
      },
      "PollingStationInvestigation": {
        "type": "object",
        "required": [
          "polling_station_id",
          "reason"
        ],
        "properties": {
          "corrected_results": {
            "type": "boolean"
          },
          "findings": {
            "type": "string"
          },
          "polling_station_id": {
            "$ref": "#/components/schemas/PollingStationId"
          },
          "reason": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "cookie_auth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "ABACUS_SESSION"
      }
    }
  }
}